Split user controller into 2 controllers

+++ src/modules/users/controllers/users-files.controller.ts
@@ -1,13 +1,11 @@
-import type { RequestWithUser } from '@types';
-import type { Request } from 'express';
+import type { I18nTranslations, RequestWithUser } from '@types';
 
 import {
-	Body,
+	BadRequestException,
 	Controller,
 	Delete,
 	Get,
 	Param,
-	Patch,
 	Post,
 	Req,
 	StreamableFile,
@@ -18,94 +16,29 @@ import {
 } from '@nestjs/common';
 import { AuthGuard } from '@nestjs/passport';
 import { FileInterceptor } from '@nestjs/platform-express';
-import {
-	ApiBearerAuth,
-	ApiBody,
-	ApiConsumes,
-	ApiOkResponse,
-	ApiOperation,
-	ApiTags,
-	ApiUnauthorizedResponse,
-} from '@nestjs/swagger';
+import { ApiBearerAuth, ApiBody, ApiConsumes, ApiOperation, ApiTags, ApiUnauthorizedResponse } from '@nestjs/swagger';
+import { I18nService } from 'nestjs-i18n';
 
+import { Errors } from '@i18n';
 import { GuardPermissions } from '@modules/auth/decorators/permissions.decorator';
 import { GuardSelfOrPermissions } from '@modules/auth/decorators/self-or-perms.decorator';
-import { UserPostByAdminDTO } from '@modules/auth/dto/register.dto';
 import { PermissionGuard } from '@modules/auth/guards/permission.guard';
 import { SelfOrPermissionGuard } from '@modules/auth/guards/self-or-perms.guard';
 import { FilesService } from '@modules/files/files.service';
-import { Role } from '@modules/roles/entities/role.entity';
 
-import { UserPatchDTO } from './dto/patch.dto';
-import { UserVisibility } from './entities/user-visibility.entity';
-import { User } from './entities/user.entity';
-import { UsersService } from './users.service';
+import { User } from '../entities/user.entity';
+import { UsersService } from '../users.service';
 
 @ApiTags('Users')
 @Controller('users')
 @UseGuards(AuthGuard('jwt'))
 @ApiBearerAuth()
-export class UsersController {
-	constructor(private readonly usersService: UsersService, private readonly filesService: FilesService) {}
-
-	@Post()
-	@UseGuards(PermissionGuard)
-	@GuardPermissions('CAN_EDIT_USER')
-	@ApiOperation({ summary: 'Create a new user' })
-	@ApiOkResponse({ description: 'The created user', type: User })
-	@ApiUnauthorizedResponse({ description: 'Insufficient permission' })
-	async create(@Body() input: UserPostByAdminDTO) {
-		return this.usersService.registerByAdmin(input);
-	}
-
-	@Patch()
-	@UseGuards(SelfOrPermissionGuard)
-	@GuardSelfOrPermissions('id', ['CAN_EDIT_USER'])
-	@ApiOperation({ summary: 'Update an existing user' })
-	@ApiOkResponse({ description: 'The updated user', type: User })
-	@ApiUnauthorizedResponse({ description: 'Insufficient permission' })
-	async update(@Body() input: UserPatchDTO, @Req() req: Request & { user: User }) {
-		return this.usersService.update(req.user.id, input);
-	}
-
-	@Delete(':id')
-	@UseGuards(SelfOrPermissionGuard)
-	@GuardSelfOrPermissions('id', ['CAN_DELETE_USER'])
-	@ApiOperation({ summary: 'Delete a user' })
-	@ApiUnauthorizedResponse({ description: 'Insufficient permission' })
-	async delete(@Param('id') id: number) {
-		return this.usersService.delete(id);
-	}
-
-	@Get(':id')
-	@UseGuards(SelfOrPermissionGuard)
-	@GuardSelfOrPermissions('id', ['CAN_READ_USER'])
-	@ApiOperation({ summary: 'Get public information of a user' })
-	@ApiOkResponse({ description: 'User data, excepted privates fields (set in the visibility table)', type: User })
-	@ApiUnauthorizedResponse({ description: 'Insufficient permission' })
-	async get(@Param('id') id: number) {
-		return this.usersService.findOne({ id });
-	}
-
-	@Get(':id/private')
-	@UseGuards(SelfOrPermissionGuard)
-	@GuardSelfOrPermissions('id', ['CAN_READ_USER_PRIVATE'])
-	@ApiOperation({ summary: 'Get information of a user' })
-	@ApiOkResponse({ description: 'User data', type: User })
-	@ApiUnauthorizedResponse({ description: 'Insufficient permission' })
-	async getAsPrivate(@Param('id') id: number) {
-		return this.usersService.findOne({ id }, false);
-	}
-
-	@Get(':id/visibility')
-	@UseGuards(SelfOrPermissionGuard)
-	@GuardSelfOrPermissions('id', ['CAN_READ_USER_PRIVATE'])
-	@ApiOperation({ summary: 'Get visibility of a user' })
-	@ApiOkResponse({ description: 'User data', type: UserVisibility })
-	@ApiUnauthorizedResponse({ description: 'Insufficient permission' })
-	async getVisibility(@Param('id') id: number) {
-		return this.usersService.findVisibility({ id });
-	}
+export class UsersFilesController {
+	constructor(
+		private readonly usersService: UsersService,
+		private readonly filesService: FilesService,
+		private readonly i18n: I18nService<I18nTranslations>,
+	) {}
 
 	@Post(':id/picture')
 	@UseGuards(SelfOrPermissionGuard)
@@ -126,6 +59,11 @@ export class UsersController {
 	})
 	@UseInterceptors(FileInterceptor('file'))
 	async editPicture(@UploadedFile() file: Express.Multer.File, @Param('id') id: number) {
+		if (!file) throw new BadRequestException(Errors.File.NotProvided({ i18n: this.i18n }));
+
+		if (typeof id !== 'number' && parseInt(id, 10) != id)
+			throw new BadRequestException(Errors.Generic.FieldInvalid({ i18n: this.i18n, type: Number, field: 'id' }));
+
 		return this.usersService.updatePicture(id, file);
 	}
 
@@ -135,6 +73,9 @@ export class UsersController {
 	@ApiOperation({ summary: 'Delete user profile picture' })
 	@ApiUnauthorizedResponse({ description: 'Insufficient permission' })
 	async deletePicture(@Param('id') id: number) {
+		if (typeof id !== 'number' && parseInt(id, 10) != id)
+			throw new BadRequestException(Errors.Generic.FieldInvalid({ i18n: this.i18n, type: Number, field: 'id' }));
+
 		return this.usersService.deletePicture(id);
 	}
 
@@ -144,6 +85,9 @@ export class UsersController {
 	@ApiOperation({ summary: 'Get user profile picture' })
 	@ApiUnauthorizedResponse({ description: 'Insufficient permission' })
 	async getPicture(@Req() req: RequestWithUser, @Param('id') id: number) {
+		if (typeof id !== 'number' && parseInt(id, 10) != id)
+			throw new BadRequestException(Errors.Generic.FieldInvalid({ i18n: this.i18n, type: Number, field: 'id' }));
+
 		const picture = await this.usersService.getPicture(id);
 
 		if (await this.filesService.canReadFile(picture, req.user as User))
@@ -171,6 +115,11 @@ export class UsersController {
 	})
 	@UseInterceptors(FileInterceptor('file'))
 	async editBanner(@UploadedFile() file: Express.Multer.File, @Param('id') id: number) {
+		if (!file) throw new BadRequestException(Errors.File.NotProvided({ i18n: this.i18n }));
+
+		if (typeof id !== 'number' && parseInt(id, 10) != id)
+			throw new BadRequestException(Errors.Generic.FieldInvalid({ i18n: this.i18n, type: Number, field: 'id' }));
+
 		return this.usersService.updateBanner(id, file);
 	}
 
@@ -180,6 +129,9 @@ export class UsersController {
 	@ApiOperation({ summary: 'Delete user profile banner' })
 	@ApiUnauthorizedResponse({ description: 'Insufficient permission' })
 	async deleteBanner(@Param('id') id: number) {
+		if (typeof id !== 'number' && parseInt(id, 10) != id)
+			throw new BadRequestException(Errors.Generic.FieldInvalid({ i18n: this.i18n, type: Number, field: 'id' }));
+
 		return this.usersService.deleteBanner(id);
 	}
 
@@ -189,21 +141,14 @@ export class UsersController {
 	@ApiOperation({ summary: 'Get user profile banner' })
 	@ApiUnauthorizedResponse({ description: 'Insufficient permission' })
 	async getBanner(@Req() req: RequestWithUser, @Param('id') id: number) {
+		if (typeof id !== 'number' && parseInt(id, 10) != id)
+			throw new BadRequestException(Errors.Generic.FieldInvalid({ i18n: this.i18n, type: Number, field: 'id' }));
+
 		const banner = await this.usersService.getBanner(id);
 
 		if (await this.filesService.canReadFile(banner, req.user as User))
 			return new StreamableFile(this.filesService.toReadable(banner));
 
 		throw new UnauthorizedException();
-	}
-
-	@Get(':id/roles')
-	@UseGuards(PermissionGuard)
-	@GuardSelfOrPermissions('id', ['CAN_READ_USER'])
-	@ApiOperation({ summary: 'Get roles of a user' })
-	@ApiOkResponse({ description: 'Roles of the user', type: [Role] })
-	@ApiUnauthorizedResponse({ description: 'Insufficient permission' })
-	async getRoleUsers(@Param('id') id: number) {
-		return this.usersService.getUserRoles(id, { show_expired: true, show_revoked: true });
 	}
 }
